mkaestatic: add compact code to support mistune v3
Module mistune again changed the API. They changed signature of
HTMLRenderer functions and dropped AstRenderer.

The number of parameters is the same for link() and image() functions so
it doesn't fails when defining our wrappers, instead it fails when they
are called with the following error:

    TypeError: get_markdown_renderer.<locals>.MyRenderer.link() got an unexpected keyword argument 'url'

As we know that AstRenderer doesn't exist in v3 we can use it the same
way to get AttributeError exception.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>